+++ Assets/Scripts/Eun/Monster_Manager.cs
@@ -197,6 +197,7 @@ public class Monster_Manager : MonoBehaviour
                 break;
 
         }
+        SaveAndLoadManager.Instance.SaveGame();
 
         SetWave();
     }
