--- Assets/Scripts/KHR/Save.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 05b079d9da4f4a14eb515724d7573b47
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
