+++ Assets/Scripts/KHR/SaveAndLoadManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4412ea1fa029df8499156b389704265c
+guid: 373e0431363971c479874c5d0eb24885
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
